Extract RunDropZone as shared component for submit and profile pages
Both pages now use the same drop zone with Overwolf + Choose Files
buttons, progress bar, and OS file paths. 1:1 identical UI.
