+++ frontend/app/leaderboards/submit/SubmitRunClient.tsx
@@ -8,7 +8,7 @@ import { useLanguage } from "@/app/contexts/LanguageContext";
 import { useAuth } from "@/app/contexts/AuthContext";
 import { t } from "@/lib/ui-translations";
 import { IS_BETA } from "@/lib/seo";
-import RunFileHelp from "@/app/components/RunFileHelp";
+import RunDropZone from "@/app/components/RunDropZone";
 
 const API = process.env.NEXT_PUBLIC_API_URL ?? "http://localhost:8000";
 
@@ -46,15 +46,6 @@ export default function SubmitRunClient() {
     dupes: number;
     errors: number;
   } | null>(null);
-  const [isDragging, setIsDragging] = useState(false);
-  const dragCounter = useRef(0);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  // Belt-and-suspenders against any source of double-fire on the upload
-  // path (the dropzone-vs-document race the native stopImmediatePropagation
-  // kills above, plus button double-click, retry, etc.). A ref instead of
-  // useState so the guard sees the latest value without waiting for a
-  // re-render — back-to-back drops fire the second handler before React
-  // commits any state from the first.
   const uploadInFlight = useRef(false);
 
   function isValidRunFile(data: any): boolean {
@@ -219,68 +210,6 @@ export default function SubmitRunClient() {
     if (user) fetchRuns();
   }, [user, fetchRuns]);
 
-  // Drag-and-drop handlers for the upload area
-  const handleDragEnter = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    dragCounter.current++;
-    setIsDragging(true);
-  }, []);
-
-  const handleDragOver = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-  }, []);
-
-  const handleDragLeave = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    dragCounter.current--;
-    if (dragCounter.current === 0) {
-      setIsDragging(false);
-    }
-  }, []);
-
-  const handleDrop = useCallback(
-    (e: React.DragEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      // The page-level `document.addEventListener("drop", …)` below also
-      // calls handleFileUpload so drops anywhere on the page are captured.
-      // React's `e.stopPropagation()` only stops further synthetic handlers
-      // — the underlying native event still bubbles past React's root to
-      // the document listener, which fires handleFileUpload a second time.
-      // Each file then gets POSTed twice: the first request inserts, the
-      // second comes back marked `duplicate`, and the UI shows "0 submitted,
-      // 2 skipped" for what was actually a clean two-file upload.
-      e.nativeEvent.stopImmediatePropagation();
-      dragCounter.current = 0;
-      setIsDragging(false);
-      if (e.dataTransfer.files && e.dataTransfer.files.length > 0) {
-        handleFileUpload(e.dataTransfer.files);
-      }
-    },
-    [handleFileUpload]
-  );
-
-  // Page-level drop handling so drops work anywhere on the page
-  useEffect(() => {
-    const handleDocDragOver = (e: DragEvent) => {
-      e.preventDefault();
-    };
-    const handleDocDrop = (e: DragEvent) => {
-      e.preventDefault();
-      if (e.dataTransfer?.files && e.dataTransfer.files.length > 0) {
-        handleFileUpload(e.dataTransfer.files);
-      }
-    };
-    document.addEventListener("dragover", handleDocDragOver);
-    document.addEventListener("drop", handleDocDrop);
-    return () => {
-      document.removeEventListener("dragover", handleDocDragOver);
-      document.removeEventListener("drop", handleDocDrop);
-    };
-  }, [handleFileUpload]);
 
   // Run submissions are server-side rejected on beta (the backend returns
   // 403 with "Submit to spire-codex.com instead"). The Navbar already
@@ -363,102 +292,7 @@ export default function SubmitRunClient() {
         />
       )}
 
-      {/* Drop zone */}
-      <div
-        onDragEnter={handleDragEnter}
-        onDragOver={handleDragOver}
-        onDragLeave={handleDragLeave}
-        onDrop={handleDrop}
-        onClick={() => fileInputRef.current?.click()}
-        className={`border-2 border-dashed rounded-lg p-6 sm:p-8 text-center cursor-pointer transition-colors ${
-          isDragging
-            ? "border-[var(--accent-gold)] bg-[var(--accent-gold)]/5"
-            : "border-[var(--border-subtle)] hover:border-[var(--border-accent)]"
-        }`}
-      >
-        <input
-          ref={fileInputRef}
-          type="file"
-          multiple
-          accept=".run,.json"
-          className="hidden"
-          onChange={(e) => {
-            if (e.target.files) handleFileUpload(e.target.files);
-            e.target.value = "";
-          }}
-        />
-        {isDragging ? (
-          <p className="text-[var(--text-primary)] font-medium">
-            {t("Drop files here...", lang)}
-          </p>
-        ) : (
-          <p className="text-[var(--text-primary)] font-medium mb-3">
-            Drop .run files here or click to browse
-          </p>
-        )}
-
-        {!isDragging && (
-          <div className="flex flex-col sm:flex-row sm:items-center sm:justify-center gap-2 sm:gap-3" onClick={(e) => e.stopPropagation()}>
-            <a
-              href="https://www.overwolf.com/app/ptrlrd-spire_codex"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="inline-flex items-center justify-center w-full sm:w-auto px-5 py-2.5 sm:py-2 rounded-lg text-sm font-medium bg-[var(--accent-gold)] text-white hover:opacity-90 transition-opacity"
-            >
-              Download Overwolf Companion App
-            </a>
-            <label className="inline-flex items-center justify-center w-full sm:w-auto px-5 py-2.5 sm:py-2 rounded-lg text-sm font-medium bg-[var(--bg-primary)] text-[var(--text-primary)] border border-[var(--border-accent)] hover:bg-[var(--bg-card-hover)] transition-colors cursor-pointer">
-              {t("Choose Files", lang)}
-              <input
-                type="file"
-                multiple
-                accept=".run,.json"
-                className="hidden"
-                onChange={(e) => {
-                  if (e.target.files) handleFileUpload(e.target.files);
-                  e.target.value = "";
-                }}
-              />
-            </label>
-          </div>
-        )}
-
-        {uploadProgress && (
-          <div className="mt-4">
-            <div className="w-full bg-[var(--bg-primary)] rounded-full h-2 mb-2">
-              <div
-                className="h-2 rounded-full bg-[var(--accent-gold)] transition-all"
-                style={{
-                  width: `${(uploadProgress.done / uploadProgress.total) * 100}%`,
-                }}
-              />
-            </div>
-            <p className="text-xs text-[var(--text-muted)]">
-              {uploadProgress.done === uploadProgress.total ? (
-                <>
-                  {t("Done!", lang)}{" "}
-                  {uploadProgress.total - uploadProgress.dupes - uploadProgress.errors}{" "}
-                  {t("submitted", lang)}
-                  {uploadProgress.dupes > 0 && (
-                    <>, {uploadProgress.dupes} {t("duplicates skipped", lang)}</>
-                  )}
-                  {uploadProgress.errors > 0 && (
-                    <>, {uploadProgress.errors} {t("invalid", lang)}</>
-                  )}
-                </>
-              ) : (
-                <>
-                  {t("Processing", lang)} {uploadProgress.done} {t("of", lang)}{" "}
-                  {uploadProgress.total}...
-                </>
-              )}
-            </p>
-          </div>
-        )}
-
-        <hr className="border-[var(--border-subtle)] my-4" />
-        <RunFileHelp />
-      </div>
+      <RunDropZone onFiles={handleFileUpload} uploadProgress={uploadProgress} />
 
       {error && (
         <p className="text-[var(--color-ironclad)] text-sm">{error}</p>
